+++ baybe/exceptions.py
@@ -33,6 +33,10 @@ class SearchSpaceMatchWarning(UserWarning):
         return self.message
 
 
+class MinimumCardinalityViolatedWarning(UserWarning):
+    """Minimum cardinality constraints are violated."""
+
+
 ##### Exceptions #####
 
 
@@ -61,6 +65,10 @@ class IncompatibleExplainerError(IncompatibilityError):
 
 class IncompatibleArgumentError(IncompatibilityError):
     """An incompatible argument was passed to a callable."""
+
+
+class InfeasibilityError(Exception):
+    """An optimization problem has no feasible solution."""
 
 
 class NotEnoughPointsLeftError(Exception):